refactor(accordian-page): turned raced page accordian tools into generics
