+++ src/shared/components/generic/index.ts
@@ -1,0 +1,8 @@
+// Core generic components
+export { GenericAccordionCard } from "./GenericAccordionCard";
+export { EntityAvatar } from "./EntityAvatar";
+export { CategoryBadge } from "./CategoryBadge";
+
+// Re-export types for convenience
+export type { EntityType, AvatarSize } from "./EntityAvatar";
+export type { BadgeSize } from "./CategoryBadge";
